+++ web/apps/photos/src/services/upload/thumbnail.ts
@@ -1,6 +1,7 @@
 import { FILE_TYPE, type FileTypeInfo } from "@/media/file-type";
 import { heicToJPEG } from "@/media/heic-convert";
 import { scaledImageDimensions } from "@/media/image";
+import * as ffmpeg from "@/new/photos/services/ffmpeg";
 import {
     toDataOrPathOrZipEntry,
     type DesktopUploadItem,
@@ -9,7 +10,6 @@ import log from "@/next/log";
 import { type Electron } from "@/next/types/ipc";
 import { ensure } from "@/utils/ensure";
 import { withTimeout } from "@/utils/promise";
-import * as ffmpeg from "services/ffmpeg";
 
 /** Maximum width or height of the generated thumbnail */
 const maxThumbnailDimension = 720;
